requirements.txt
django-money
phonenumbers
django-phonenumber-field
Django